Do not use osxtime_t for available times
Previously, `osxtime_t` was used to set and retrieve available times, as
well as storing `LMIC.baseAvail`. This allowed the code to not worry
about timer overflows, at the expense of some extra complexity
converting between time and xtime and the extra code size of handling
64-bit variables.

Because in practice other parts of BasicMac did still use 32-bit
timestamps for calculations, the available times could never end up more
than 2^31 ticks in the future anyway. Given the maximum airtime, 1%
duty cycle and minimum 15.5μ ticks, this also gives sufficient range.

So this commit lets available time handling use 32-bit `ostime_t`
everywhere. This removes the need to convert between different sizes
timestamps. but does need to modify comparisons into subtractions and
casts to `ostimediff_t` to handle overflow properly.

This also adds a job that runs every now and then to update
`LMIC.baseAvail` and prevent available times from ending up too far in
the past (and wrapping into the future) when a given channel or band is
not used for a longer period of time.

In one compiled example, this commit saves around 200 bytes of flash,
and adds 8 bytes of RAM (the smaller gain of the timestamps is cancelled
by the extra job). On AVR, which needs a lot more instructions to handle
64-bit integers and is worse at optimizing them, this saves nearly 1000
bytes of flash at the expense of 6 bytes of RAM.